requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
requests
djoser
secrets
scapy
fastapi
uvicorn
